Split running an express app from creating one
